Convert to using jest rather than karma/jasmine
