+++ .eslintrc
@@ -2,26 +2,36 @@
   "env": {
     "commonjs": true,
     "es6": true,
-    "jasmine": true
+    "jest/globals": true
   },
   "parserOptions": {
     "ecmaVersion": 2018,
     "sourceType": "module"
   },
   "plugins": [
-    "jasmine"
+    "jest"
   ],
-  "extends": "eslint:recommended",
+  "extends": ["eslint:recommended", "plugin:jest/recommended"],
   "rules": {
+    "array-bracket-spacing": ["error", "never"],
+    "brace-style": ["error","1tbs", {}],
     "comma-dangle": [2,"never"],
+    "curly": "error",
     "linebreak-style": ["error", "unix"],
     "no-console": "off",
     "no-extra-parens": [2,"all"],
     "no-use-before-define": 2,
-    "brace-style": [1,"1tbs",{}],
     "no-mixed-spaces-and-tabs": 2,
+    "object-curly-spacing": ["error", "never"],
     "quotes": ["error", "single"],
-    "semi": ["error", "always"]
+    "semi": ["error", "always"],
+    "space-before-blocks": ["error", "always"],
+    "space-before-function-paren": ["error", "never"],
+    "jest/no-disabled-tests": "warn",
+    "jest/no-focused-tests": "error",
+    "jest/no-identical-title": "error",
+    "jest/prefer-to-have-length": "warn",
+    "jest/valid-expect": "error"
   },
   "globals": {
     "window": true,
