+++ buble.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+    objectAssign: 'Object.assign',
+    transforms: {
+        dangerousForOf: true
+    }
+};
