+++ jest.config.js
@@ -1,0 +1,15 @@
+module.exports = {
+    verbose: true,
+    transform: {
+        '^.+\\.js$': 'buble-jest'
+    },
+    collectCoverageFrom: [
+        'src/js/**/*.js',
+        '!*.js',
+        '!**/node_modules/**',
+        '!**/dist/**',
+        '!**/coverage/**'
+    ],
+    clearMocks: true,
+    restoreMocks: true
+}
