+++ package.json
@@ -14,8 +14,9 @@
     "prepare": "npm run build",
     "serve": "live-server --port=9000 --cors dist",
     "start": "run-p watch serve",
-    "test": "karma start",
-    "test:watch": "karma start --debug --no-single-run --browsers Chrome",
+    "test": "jest --collectCoverage=true",
+    "test:debug": "node --inspect-brk node_modules/.bin/jest --runInBand",
+    "test:watch": "jest --watch",
     "watch": "run-p watch:*",
     "watch:css": "nodemon -w src/stylesheets -e scss -x \"npm run build:css\"",
     "watch:fonts": "npm run build:fonts",
@@ -59,15 +60,11 @@
     "@rollup/plugin-node-resolve": "^11.0.0",
     "ajv": "^6.12.6",
     "autoprefixer": "^10.0.4",
+    "buble": "^0.20.0",
+    "buble-jest": "^1.0.1",
     "eslint": "^7.14.0",
-    "eslint-plugin-jasmine": "^4.1.1",
-    "jasmine-core": "^3.6.0",
-    "karma": "^5.2.3",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-firefox-launcher": "^2.1.0",
-    "karma-jasmine": "^4.0.1",
-    "karma-js-coverage": "^0.4.0",
-    "karma-rollup-preprocessor": "^7.0.5",
+    "eslint-plugin-jest": "^24.1.3",
+    "jest": "^26.6.3",
     "live-server": "^1.2.1",
     "nodemon": "^2.0.6",
     "npm-run-all": "^4.1.5",
@@ -78,7 +75,6 @@
     "postcss-csso": "^5.0.0",
     "postcss-flexbugs-fixes": "^5.0.2",
     "rollup": "^2.34.0",
-    "rollup-plugin-istanbul": "^2.0.1",
     "sass": "^1.29.0",
     "stylelint": "^13.8.0",
     "stylelint-config-recommended-scss": "^4.2.0",
